Show isLastOwner property on listRoomsOfUser endpoint
