+++ app/lib/server/methods/getRoomRoles.js
@@ -1,9 +1,9 @@
 import { Meteor } from 'meteor/meteor';
 import { check } from 'meteor/check';
-import _ from 'underscore';
 
 import { settings } from '../../../settings';
-import { Subscriptions, Users, Roles } from '../../../models';
+import { getRoomRoles } from '../lib/getRoomRoles';
+
 
 Meteor.methods({
 	getRoomRoles(rid) {
@@ -15,29 +15,6 @@ Meteor.methods({
 
 		check(rid, String);
 
-		const options = {
-			sort: {
-				'u.username': 1,
-			},
-			fields: {
-				rid: 1,
-				u: 1,
-				roles: 1,
-			},
-		};
-
-		const UI_Use_Real_Name = settings.get('UI_Use_Real_Name') === true;
-
-		const roles = Roles.find({ scope: 'Subscriptions', description: { $exists: 1, $ne: '' } }).fetch();
-		const subscriptions = Subscriptions.findByRoomIdAndRoles(rid, _.pluck(roles, '_id'), options).fetch();
-
-		if (!UI_Use_Real_Name) {
-			return subscriptions;
-		}
-		return subscriptions.map((subscription) => {
-			const user = Users.findOneById(subscription.u._id);
-			subscription.u.name = user && user.name;
-			return subscription;
-		});
+		return getRoomRoles(rid);
 	},
 });
